Moving config to Core
Movilg the cofig to core folder / namespace since it is only used in there. no reason to be in the App namespace
