+++ Core/Container.php
@@ -2,7 +2,7 @@
 namespace Core;
 
 use PDO;
-use App\Config;
+
 
 class Container{
