+++ Core/Error.php
@@ -53,7 +53,7 @@ class Error
         http_response_code($code);
 
         //Constructing the error message to send to twig
-        if (\App\Config::SHOW_ERRORS) {
+        if (Config::SHOW_ERRORS) {
             $viewData['showErrors'] = true; //sending the config option down to twig
             $viewData['classException'] = get_class($exception);
             $viewData['stackTrace'] = $exception->getTraceAsString();
